rename TeaserSection components according to Page.vue: ProductsSection
