+++ presenter/src/components/sections/ProductsSection.stories.ts
@@ -1,18 +1,18 @@
 import { SBComp } from '#types/SBComp'
 
-import ThirdSection from './ThirdSection.vue'
+import ProductsSection from './ProductsSection.vue'
 
 import type { Meta, StoryObj } from '@storybook/vue3'
 
 // More on how to set up stories at: https://storybook.js.org/docs/vue/writing-stories/introduction
 const meta = {
   title: 'Sections/Section3',
-  component: ThirdSection as SBComp,
+  component: ProductsSection as SBComp,
   // This component will have an automatically generated docsPage entry: https://storybook.js.org/docs/vue/writing-docs/autodocs
   tags: ['autodocs'],
   argTypes: {},
   args: {}, // default value
-} satisfies Meta<typeof ThirdSection>
+} satisfies Meta<typeof ProductsSection>
 
 export default meta
 type Story = StoryObj<typeof meta>
