+++ presenter/src/components/sections/ProductsSection.vue
